Dork-Backend/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
anthropic==0.34.1
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
